backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-openai
python-dotenv
pypdf
faiss-cpu
sentence-transformers
